Split widget: glass pill
