+++ extension/devtools/app/lib/main.dart
@@ -23,6 +23,7 @@ part 'widgets/action_pill.dart';
 part 'widgets/filter_chip.dart';
 part 'widgets/glass_card.dart';
 part 'widgets/glass_input.dart';
+part 'widgets/glass_pill.dart';
 part 'widgets/panel_placeholder.dart';
 part 'widgets/panel_scroll_view.dart';
 part 'widgets/sort_header_cell.dart';
@@ -3713,49 +3714,6 @@ class _InfoRow extends StatelessWidget {
           SizedBox(width: 84, child: Text(label, style: textTheme.bodySmall)),
           Expanded(child: Text(value, style: textTheme.bodyMedium)),
         ],
-      ),
-    );
-  }
-}
-
-class _GlassPill extends StatelessWidget {
-  const _GlassPill({required this.child, this.padding, this.color});
-
-  final Widget child;
-  final EdgeInsets? padding;
-  final Color? color;
-
-  @override
-  Widget build(BuildContext context) {
-    final brightness = Theme.of(context).brightness;
-    final tint =
-        color ??
-        (brightness == Brightness.dark
-            ? Colors.white.withValues(alpha: 0.08)
-            : Colors.white.withValues(alpha: 0.9));
-
-    return ClipRRect(
-      borderRadius: BorderRadius.circular(999),
-      child: BackdropFilter(
-        filter: ImageFilter.blur(sigmaX: 12, sigmaY: 12),
-        child: Container(
-          padding:
-              padding ??
-              const EdgeInsets.symmetric(horizontal: 12, vertical: 8),
-          decoration: BoxDecoration(
-            color: tint,
-            borderRadius: BorderRadius.circular(999),
-            border: Border.all(
-              color: brightness == Brightness.dark
-                  ? Colors.white.withValues(alpha: 0.1)
-                  : Colors.white.withValues(alpha: 0.5),
-            ),
-          ),
-          child: DefaultTextStyle.merge(
-            style: Theme.of(context).textTheme.labelMedium,
-            child: child,
-          ),
-        ),
       ),
     );
   }
